mi-app-python10/backend/requirements.txt
Flask
neo4j
python-dotenv
psycopg2-binary
Flask-SQLAlchemy
Werkzeug